Better errors in channel; moved in separate mod
