+++ relayer/src/channel.rs
@@ -5,7 +5,6 @@ use std::time::Duration;
 use anomaly::BoxError;
 use prost_types::Any;
 use serde::Serialize;
-use thiserror::Error;
 use tracing::{debug, error, info, warn};
 
 use ibc::events::IbcEvent;
@@ -16,9 +15,7 @@ use ibc::ics04_channel::msgs::chan_open_ack::MsgChannelOpenAck;
 use ibc::ics04_channel::msgs::chan_open_confirm::MsgChannelOpenConfirm;
 use ibc::ics04_channel::msgs::chan_open_init::MsgChannelOpenInit;
 use ibc::ics04_channel::msgs::chan_open_try::MsgChannelOpenTry;
-use ibc::ics24_host::identifier::{
-    ChainId, ChannelId, ClientId, ConnectionId, PortChannelId, PortId,
-};
+use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId, PortId};
 use ibc::tx_msg::Msg;
 use ibc::Height;
 use ibc_proto::ibc::core::channel::v1::QueryConnectionChannelsRequest;
@@ -26,12 +23,14 @@ use ibc_proto::ibc::core::channel::v1::QueryConnectionChannelsRequest;
 use crate::chain::counterparty::{channel_connection_client, channel_state_on_destination};
 use crate::chain::handle::ChainHandle;
 use crate::connection::Connection;
-use crate::error::Error;
-use crate::foreign_client::{ForeignClient, ForeignClientError};
+use crate::foreign_client::ForeignClient;
 use crate::object::Channel as WorkerChannelObject;
 use crate::supervisor::Error as WorkerChannelError;
 use crate::util::retry::RetryResult;
 use crate::util::retry::{retry_count, retry_with_index};
+
+mod error;
+pub use error::ChannelError;
 
 mod retry_strategy {
     use std::time::Duration;
@@ -48,44 +47,6 @@ mod retry_strategy {
     pub fn default() -> impl Iterator<Item = Duration> {
         clamp_total(Fibonacci::from(INITIAL_DELAY), MAX_DELAY, MAX_TOTAL_DELAY)
     }
-}
-
-#[derive(Debug, Error)]
-pub enum ChannelError {
-    #[error("failed with underlying cause: {0}")]
-    Failed(String),
-
-    #[error("failed due to missing local channel id")]
-    MissingLocalChannelId,
-
-    #[error("failed due to missing counterparty channel id")]
-    MissingCounterpartyChannelId,
-
-    #[error("failed due to missing counterparty connection")]
-    MissingCounterpartyConnection,
-
-    #[error("channel constructor failed due to missing connection id on chain id {0}")]
-    MissingLocalConnection(ChainId),
-
-    #[error("failed during an operation on client ({0}) hosted by chain ({1}) with error: {2}")]
-    ClientOperation(ClientId, ChainId, ForeignClientError),
-
-    #[error("failed during a query to chain id {0} with underlying error: {1}")]
-    QueryError(ChainId, Error),
-
-    #[error(
-        "failed during a transaction submission step to chain id {0} with underlying error: {1}"
-    )]
-    SubmitError(ChainId, Error),
-
-    #[error("the channel is partially open ({0}, {1})")]
-    PartialOpenHandshake(State, State),
-
-    #[error("channel {0} on chain {1} has no counterparty channel id")]
-    IncompleteChannelState(PortChannelId, ChainId),
-
-    #[error("channel {0} on chain {1} expected to have counterparty {2} (but instead has {3})")]
-    MismatchingChannelEnds(PortChannelId, ChainId, PortChannelId, PortChannelId),
 }
 
 #[derive(Clone, Debug, Serialize)]
@@ -464,21 +425,25 @@ impl Channel {
             let a_channel = channel
                 .src_chain()
                 .query_channel(&channel.src_port_id(), src_channel_id, Height::zero())
-                .map_err(|_| {
-                    ChannelError::Failed(format!(
-                        "failed to query source chain {}",
-                        channel.src_chain().id()
-                    ))
+                .map_err(|e| {
+                    ChannelError::HandshakeFinalize(
+                        channel.src_port_id().clone(),
+                        src_channel_id.clone(),
+                        channel.src_chain().id(),
+                        e.to_string(),
+                    )
                 })?;
 
             let b_channel = channel
                 .dst_chain()
                 .query_channel(&channel.dst_port_id(), dst_channel_id, Height::zero())
-                .map_err(|_| {
-                    ChannelError::Failed(format!(
-                        "failed to query destination chain {}",
-                        channel.dst_chain().id()
-                    ))
+                .map_err(|e| {
+                    ChannelError::HandshakeFinalize(
+                        channel.dst_port_id().clone(),
+                        dst_channel_id.clone(),
+                        channel.dst_chain().id(),
+                        e.to_string(),
+                    )
                 })?;
 
             Ok((*a_channel.state(), *b_channel.state()))
